requirements.txt
paho-mqtt
pika
coapthon3
scapy
pandas
matplotlib
